+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [2.16.0]
+
+### Added
+
+* [react-native-videoeditorsdk] Added duration action for text and stickers.
+* [react-native-videoeditorsdk] Added `VideoEditorResult.segments`, `VideoEditorResult.videoSize`, and `VideoEditorResult.release()` which enable serialization of the individual video composition components if `configuration.export.video.segments` is enabled.
+
 ## [2.15.0]
 
 ### Changed